requirements.txt
streamlit>=1.35.0
google-generativeai>=0.3.2
python-dotenv>=1.0.1
